fixed representation of open policies
If there are no restrictions - i.e. an empty list on all constraints -
previously only the header row was shown. Now a row with a blank
user/role cell is being displayed.

This was primarily required because the interactive row clones the last
entry in the table - but it seems sensible in general.